feat: 🎉 Create ProfilePage component to display GitHub user profile and repositories with SSR support
